Move instance and hub code to generic zone code
