--- src/map/systems/hub/mod.rs
@@ -1,5 +1,0 @@
-pub mod generate_hub_layout;
-pub mod spawn_hub_entities;
-
-pub use generate_hub_layout::insert_hub_layout;
-pub use spawn_hub_entities::spawn_map_entities;
