+++ src/map/systems/load_instance_data.rs
