+++ src/map/systems/zone/mod.rs
@@ -1,11 +1,11 @@
 pub mod background;
 pub mod finish_create_zone;
-pub mod load_instance_data;
-pub mod render_tilemap;
 pub mod spawn_zone_colliders;
+pub mod spawn_zone_entities;
+pub mod spawn_zone_tilemap;
 
 pub use background::*;
 pub use finish_create_zone::finish_create_zone;
-pub use load_instance_data::*;
-pub use render_tilemap::render_tilemap;
 pub use spawn_zone_colliders::spawn_zone_colliders;
+pub use spawn_zone_entities::spawn_zone_entities;
+pub use spawn_zone_tilemap::spawn_zone_tilemap;
